+++ Components/Core/Conditions/Categories/PlayerAttributes_config.lua
@@ -22,9 +22,6 @@ local strlowerCache = TMW.strlowerCache
 local GetSpellInfo = TMW.GetSpellInfo
 local GetSpellName = TMW.GetSpellName
 
-local GetNumTrackingTypes = GetNumTrackingTypes or C_Minimap.GetNumTrackingTypes
-local GetTrackingInfo = GetTrackingInfo or C_Minimap.GetTrackingInfo
-
 local _, pclass = UnitClass("Player")
 
 
@@ -152,7 +149,34 @@ Module.showColorHelp = false
 Module.helpText = L["SUG_TOOLTIPTITLE_GENERIC"]
 
 local TrackingCache = {}
-if GetNumTrackingTypes and GetTrackingInfo and GetNumTrackingTypes() > 0 then
+if C_Minimap and C_Minimap.GetTrackingInfo then
+	local GetTrackingInfo = C_Minimap.GetTrackingInfo
+	local GetNumTrackingTypes = C_Minimap.GetNumTrackingTypes
+	-- Wow 11.0+
+	
+	function Module:Table_Get()
+		for i = 1, GetNumTrackingTypes() do
+			local data = GetTrackingInfo(i)
+			TrackingCache[i] = strlower(data.name)
+		end
+		
+		return TrackingCache
+	end
+
+	function Module:Entry_AddToList_1(f, id)
+		local info = GetTrackingInfo(id)
+	
+		f.Name:SetText(info.name)
+		f.ID:SetText(nil)
+	
+		f.tooltiptitle = info.name
+		
+		f.insert = info.name
+	
+		f.Icon:SetTexture(info.texture)
+	end
+
+elseif GetNumTrackingTypes and GetTrackingInfo and GetNumTrackingTypes() > 0 then
 	function Module:Table_Get()
 		for i = 1, GetNumTrackingTypes() do
 			local name, _, active = GetTrackingInfo(i)