+++ Options/CHANGELOG.lua
@@ -4,7 +4,8 @@ TMW.CHANGELOG_LASTVER="7.4.0"
 
 TMW.CHANGELOG = [==[
 ## v11.0.2
-* Fix spell dragging to icons (again)
+* Fix: #2178 Spell dragging to icons broken (again) in WoW 11.0
+* Fix: #2179 Tracking Active condition broken in WoW 11.0
 
 ## v11.0.1
 * Fix: #2174 - Autocast conditions not working